+++ extensions/wrapper/wrapper-common-api/src/main/java/de/sovity/edc/ext/wrapper/api/common/model/UiPolicyConstraint.java
@@ -23,26 +23,20 @@ import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-/**
- * Opinionated subset of the EDC policy.
- *
- * @author tim.dahlmanns@isst.fraunhofer.de
- */
 @Getter
 @Setter
 @ToString
 @AllArgsConstructor
 @Builder(toBuilder = true)
 @RequiredArgsConstructor
-@Schema(description = "Type-Safe OpenAPI generator friendly Policy DTO that supports an opinionated"
-        + " subset of the original EDC Policy Entity.")
-public class PolicyDto {
+@Schema(description = "ODRL AtomicConstraint as supported by our UI")
+public class UiPolicyConstraint {
+    @Schema(description = "Left side of the expression.", requiredMode = RequiredMode.REQUIRED)
+    private String left;
 
-    @Schema(description = "Legacy JSON as built by the Management API. Will be replaced "
-            + "in the future by a type-safe variant without polymorphisms that can be used "
-            + "for our generated clients.", requiredMode = Schema.RequiredMode.REQUIRED)
-    private String legacyPolicy;
+    @Schema(description = "Operator, e.g. EQ", requiredMode = RequiredMode.REQUIRED)
+    private OperatorDto operator;
 
-    @Schema(description = "Permission for this policy", requiredMode = RequiredMode.REQUIRED)
-    private PermissionDto permission;
+    @Schema(description = "Right side of the expression", requiredMode = RequiredMode.REQUIRED)
+    private UiPolicyLiteral right;
 }